Refactor code, Add option to load/save to own file

+++ directory.rb
@@ -3,21 +3,17 @@
 def interactive_menu
   loop do
     print_menu
-    process(STDIN.gets.chomp)
+    option(STDIN.gets.chomp)
   end
 end
 
 def input_students
   puts "Please enter the names of the students"
   puts "To finish, just hit return twice"
-  # get the first name
   name = STDIN.gets.chomp
-  # while the name is not empty, repeat this code
   while !name.empty? do
-    # add the student hash to the array
     add_students(name, cohort = "November")
     puts "Now we have #{@students.count} students"
-    # get another name from the user
     name = STDIN.gets.chomp
   end
 end
@@ -30,8 +26,8 @@ end
 def print_menu
   puts "1. Input the students"
   puts "2. Show the students"
-  puts "3. Save the list to students.csv"
-  puts "4. Load the list from students.csv"
+  puts "3. Save the list to file"
+  puts "4. Load the list from file"
   puts "9. Exit"
 end
 
@@ -41,25 +37,19 @@ def show_students
     print_footer
 end
 
-def process(selection)
+def option(selection)
   case selection
-  when "1"
-    puts "You selected 'add students'"
+  when "1" then puts "You selected 'add students'"
     input_students
-  when "2"
-    puts "You selected 'show students'"
+  when "2" then puts "You selected 'show students'"
     show_students
-  when "3"
-    puts "You selected 'save students'"
+  when "3" then puts "You selected 'save students'"
     save_students
-  when "4"
-    puts "You selected 'load students'"
+  when "4" then puts "You selected 'load students'"
     load_students
-  when "9"
-    puts 'Exiting'
+  when "9" then puts 'Exiting'
     exit
-  else
-      puts "I don't know what you meant, please try again"
+  else puts "I don't know what you meant, please try again"
   end
 end
 
@@ -74,21 +64,45 @@ def print_footer
 end
 
 def save_students
-  file = File.open("students.csv", "w")
-  @students.each do |student|
+   puts "Please enter a filename or enter return to use default file"
+   filename = STDIN.gets.chomp
+   if File.exists?(filename)
+    file = File.open((filename), "w")
+    @students.each do |student|
     student_data = [student[:name], student[:cohort]]
     csv_line = student_data.join(",")
     file.puts csv_line
+    end
+    else
+    file = File.open("students.csv", "w")
+    @students.each do |student|
+    student_data = [student[:name], student[:cohort]]
+    csv_line = student_data.join(",")
+    file.puts csv_line
+    end
   end
   file.close
 end
 
 def load_students(filename = "students.csv")
-  file = File.open(filename, "r")
-  file.readlines.each do |line|
-  name, cohort = line.chomp.split(',')
-  add_students(name, cohort = "November")
-  end
+  puts "Please enter a filename or enter return to use default file"
+  filename = STDIN.gets.chomp
+    if File.exists?(filename)
+      puts "Loaded from #{filename}"
+      file = File.open(filename, "r")
+      file.readlines.each do |line|
+      name, cohort = line.chomp.split(',')
+      add_students(name, cohort = "November")
+      end
+    else
+      puts "Loaded from 'students.csv'"
+      filename = "students.csv"
+      file = File.open(filename, "r")
+      file.readlines.each do |line|
+      name, cohort = line.chomp.split(',')
+      add_students(name, cohort = "November")
+      end
+    end
   file.close
 end
